==> notebooks/2_Parse_PDF.ipynb <==
# Parse banks report pdf

---

The goal here is to extract paragraphs and relevant form of text from the pdf files located in the **data/external** folder.

 We will try our best to use them to evaluate our model.  



import sys
import os
import json
from pprint import pprint

import pandas as pd
import numpy as np

from google.colab import drive
drive.mount('/content/drive', force_remount=True)

path = "/content/drive/MyDrive/Hackathon_ISEP/Data/"
test_path = path+"test/"

!git clone https://github.com/Garage-ISEP/hackathon-ia-developpement-durable

!pip install tika

from tika import parser  

## Definition of the parser 

def parse_pdf(pdf_path, filename):
    parsed_pdf = parser.from_file(pdf_path)
    data = parsed_pdf['content'] 
    data_split = data.split("\n\n")
    result = list(filter(lambda x: len(x.split(' ')) > 10, data_split)) # Keep paragraph of +10 words
    result = list(map(lambda x: x.replace('\n', ' '), result))
    df = pd.DataFrame(result, columns=['Text'])
    if filename:
        df.to_csv(test_path+filename+'.csv')
    return result

## Parsing

'''
    For the given path, get the List of all files in the directory tree 
'''
def getListOfFiles(dirName):
    # create a list of file and sub directories 
    # names in the given directory 
    listOfFile = os.listdir(dirName)
    allFiles = list()
    # Iterate over all the entries
    for entry in listOfFile:
        # Create full path
        fullPath = os.path.join(dirName, entry)
        # If entry is a directory then get the list of files in this directory 
        if os.path.isdir(fullPath):
            allFiles = allFiles + getListOfFiles(fullPath)
        else:
            allFiles.append((fullPath, entry.split('.')[0]))
                
    return allFiles   

dirName = "/content/hackathon-ia-developpement-durable/data/external/Exemples projets Banque Mondiale"
    
# Get the list of all files in directory tree at given path
listOfFiles = getListOfFiles(dirName)

# Print the files
for elem in listOfFiles:
    print(elem)

for path, filename in listOfFiles:
    parse_pdf(path, filename)
