==> notebooks/3_Training.ipynb <==
## Prérequis

!pip install git+https://github.com/kerighan/textgraph
!pip install openpyxl
!pip install scikit-learn -U
!pip install bert_embedding
!pip install torch
!pip install git+https://github.com/kerighan/convectors

import pandas as pd
from textgraph.graph import TextGraph
import networkx as nx
import numpy as np
import regex as re

## Constitution du dataset

xls = pd.ExcelFile('data/afd_snaps_labeled_cibles.xlsx')
df12 = pd.read_excel(xls, 'SDG 12')
df15 = pd.read_excel(xls, 'SDG 15')
df16 = pd.read_excel(xls, 'SDG 16')

unlabelled_df = pd.read_csv("data/raw_filtered.csv")
def get_unlabelled(data, SDG):
    
    data = data[data["SDG"+str(SDG)] == 1].drop(columns = "Unnamed: 0").reset_index(drop = True)
    data.columns = ["Text"] + list(data.columns[1:])
    data = data[["Text"]]
    data.Text = [re.sub("[^a-zA-Z0-9]", " ", text) for text in data.Text]
    empty_cols = ["" for i in range(len(data))]
    data["Manual_1"] =empty_cols
    data["Manual_2 "] = empty_cols
    return(data)

    
unlabelled_df12 = get_unlabelled(unlabelled_df, 12)
unlabelled_df15 = get_unlabelled(unlabelled_df, 15)
unlabelled_df16 = get_unlabelled(unlabelled_df, 16)

## Aggrégation des données scrappées

unlabelled_aug = pd.read_csv("data/unlabeled_snaps.csv", sep = ";")

unlabelled_aug_df12 = get_unlabelled(unlabelled_aug, 12)
unlabelled_aug_df15 = get_unlabelled(unlabelled_aug, 15)
unlabelled_aug_df16 = get_unlabelled(unlabelled_aug, 16)

labelled_desc_12 = pd.read_csv("data/aug_cibles_12.csv").drop(columns = "Unnamed: 0")
labelled_desc_15 = pd.read_csv("data/aug_cibles_15.csv").drop(columns = "Unnamed: 0")
labelled_desc_16 = pd.read_csv("data/aug_cibles_16.csv").drop(columns = "Unnamed: 0")

full_df12 = pd.concat([df12,labelled_desc_12, unlabelled_aug_df12, unlabelled_df12], axis = 0).reset_index(drop = True)
full_df15 = pd.concat([df15,labelled_desc_15, unlabelled_aug_df15, unlabelled_df15], axis = 0).reset_index(drop = True)
full_df16 = pd.concat([df16,labelled_desc_16, unlabelled_aug_df16, unlabelled_df16], axis = 0).reset_index(drop = True)

## Calcul de l'embedding de chaque mot

import pickle
def load_wv():
    with open("wv.pkl", "rb") as f:
        wv = pickle.load(f)
    return wv

import pickle
def save_wv(wv):
    with open("wv.pkl", "wb") as f:
        pickle.dump(wv, f)

from bert_embedding import BertEmbedding

def generate_embeddings(df, first_time = False):
    df.Text = [text.replace(".","").replace(",", "").lower() for text in df.Text]
    if first_time:
        wv = {}
    else:
        wv = load_wv()
    new_words = []
    for text in df["Text"]:
        sent = text.split()
        for word in sent:
            if word in wv or word in new_words:
                continue
            new_words.append(word)
    new_words = list(set(new_words))
    bert_embedding = BertEmbedding()
    result = bert_embedding(new_words)
    for pair in result:

        wv[pair[0][0]] = pair[1][0]
    save_wv(wv)
    return wv

wv = generate_embeddings(full_df12)
print("done with 12")
wv = generate_embeddings(full_df15)
print("done with 15")
wv = generate_embeddings(full_df16)
print("done with 16")

## Création du graph

def generate_graph(df):
    G = TextGraph(wv_threshold=.5, stopwords=["en"]).fit(full_df12["Text"], wv = wv)
    A = nx.adjacency_matrix(G)
    degrees = []
    for d in G.degree(weight=None):
        if d[1] == 0:
            degrees.append(0)
        else:
            degrees.append(d[1]**(-0.5))

    print(len(degrees))
    print(A.shape)
    degrees = np.diag(degrees)
    A_hat = degrees @ A @ degrees
    return G, A_hat

###Modele avec co occurence

#G= TextGraph(wv_threshold=.5, stopwords=["en"]).fit(full_df12["Text"])
#nx.draw(G)
#A = nx.adjacency_matrix(G)

## Création du GCN

import torch
import torch.nn as nn
import torch.nn.functional as F

class gcn(nn.Module):
    def __init__(self, X_size, A_hat, num_classes, bias=True): # X_size = num features
        super(gcn, self).__init__()
        self.A_hat = torch.tensor(A_hat, requires_grad=False).float()
        self.weight = nn.parameter.Parameter(torch.FloatTensor(X_size, hidden_size_1))
        var = 2./(self.weight.size(1)+self.weight.size(0))
        self.weight.data.normal_(0,var)
        self.weight2 = nn.parameter.Parameter(torch.FloatTensor(hidden_size_1, hidden_size_2))
        var2 = 2./(self.weight2.size(1)+self.weight2.size(0))
        self.weight2.data.normal_(0,var2)
        if bias:
            self.bias = nn.parameter.Parameter(torch.FloatTensor(hidden_size_1))
            self.bias.data.normal_(0,var)
            self.bias2 = nn.parameter.Parameter(torch.FloatTensor(hidden_size_2))
            self.bias2.data.normal_(0,var2)
        else:
            self.register_parameter("bias", None)
        self.fc1 = nn.Linear(hidden_size_2, num_classes)
        
    def forward(self, X): ### 2-layer GCN architecture
        X = torch.mm(X, self.weight)
        if self.bias is not None:
            X = (X + self.bias)
        X = F.relu(torch.mm(self.A_hat, X))
        X = torch.mm(X, self.weight2)
        if self.bias2 is not None:
            X = (X + self.bias2)
        X = F.relu(torch.mm(self.A_hat, X))
        return self.fc1(X)

hidden_size_1 = 200
hidden_size_2 = 100

num_epochs=100
lr=0.011
model_no =0




## Entrainement du GCN

def load_pickle(filename):
    completeName = os.path.join("save/",\
                                filename)
    with open(completeName, 'rb') as pkl_file:
        data = pickle.load(pkl_file)
    return data

def save_as_pickle(filename, data):
    completeName = os.path.join("save/",\
                                filename)
    with open(completeName, 'wb') as output:
        pickle.dump(data, output)

def mapper(df):
    map = {}
    df2 = df.copy()
    j = 0
    for i in list(set(df["Manual_1"])):
        map[i] = j
        j+=1
    labelled = df[df["Manual_1"] != ""]
    df2.loc[:(len(labelled)-1),"Manual_1"] = [map[df["Manual_1"][i]] for i in range(len(labelled))]
    return map , df2

import torch.nn as nn
import torch.optim as optim
import os


def evaluate(output, labels_e):
    _, labels = output.max(1)
    labels = labels.numpy()
    return sum([(e) for e in labels_e] == labels)/len(labels)



def train(data):
    map, data = mapper(data)
    print(map)

    G, A_hat = generate_graph(data)

    I = np.eye(A_hat.shape[0])
    net = gcn(A_hat.shape[0], A_hat, num_classes = len(list(set(data["Manual_1"]))) )

    

    f = torch.from_numpy(I).float()
    selected = np.array(range(len(data[data["Manual_1"] != ""]))) # Indices des datas labelisées
    test_idx = np.random.choice(selected, size = 10)
    
    

    selected = np.array(list(set(selected) - set(test_idx)))

    
    labels_selected = data.loc[selected, "Manual_1"]
    labels_not_selected = data.loc[test_idx, "Manual_1"]

    criterion = nn.CrossEntropyLoss()
    optimizer = optim.Adam(net.parameters(), lr=lr)
    scheduler = optim.lr_scheduler.MultiStepLR(optimizer, milestones=[1000,2000,3000,4000,5000,6000], gamma=0.77)

    start_epoch, best_pred = 0, 0
    losses_per_epoch, evaluation_untrained = [], []

    net.train()

    evaluation_trained = []
    for e in range(start_epoch, num_epochs):
        optimizer.zero_grad()
        output = net(f)
        loss = criterion(output[selected], torch.tensor(labels_selected).long())
        losses_per_epoch.append(loss.item())
        loss.backward()
        optimizer.step()
        if e % 10 == 0:
            ### Evaluate other untrained nodes and check accuracy of labelling
            net.eval()
            with torch.no_grad():
                pred_labels = net(f)
                trained_accuracy = evaluate(output[selected], labels_selected)
                untrained_accuracy = evaluate(pred_labels[test_idxs], labels_not_selected)
            evaluation_trained.append((e, trained_accuracy))
            evaluation_untrained.append((e, untrained_accuracy))
            print("[Epoch ", e, "]: Evaluation accuracy of trained nodes:", trained_accuracy)
            print("[Epoch ", e, "]: Evaluation accuracy of test nodes:", untrained_accuracy)
            #print("Labels of trained nodes: \n", output[selected].max(1)[1])
            net.train()
            if trained_accuracy > best_pred:
                best_pred = trained_accuracy
                torch.save({
                    'epoch': e + 1,\
                    'state_dict': net.state_dict(),\
                    'best_acc': trained_accuracy,\
                    'optimizer' : optimizer.state_dict(),\
                    'scheduler' : scheduler.state_dict(),\
                }, os.path.join("save/" ,\
                "test_model_best_%d.pth.tar" % model_no))
    scheduler.step()
    net(f)
    return output.max(1)[1], map

train(full_df15)

train(full_df16)

train(full_df12)

## Testing

import os
files = os.listdir("data/test/")

df_target = pd.read_csv("data/test/processed_odd15.csv", sep = ";").drop(columns = "Unnamed: 0")
empty_cols = ["" for i in range(len(df_target))]

df_target["Manual_1"] =empty_cols
df_target["Manual_2 "] = empty_cols
df_target.columns = ["Text", "Manual_1", "Manual_2 "]

test_df15 = pd.concat([full_df15, df_target]).reset_index(drop = True)

wv = generate_embeddings(test_df15)

out, map = train(test_df15)

out

for i in range(len(df_target)):
    print(out[-i-1])

map
